Rename creation dataset and add some missing code
Article creation dataset was incorrectly named, it shows articles that
were created that are still existing, e.g. "surviving" articles (but
does not use a well-defined measure of survival). Since we might be
interested in the data later, the dataset has been renamed.

Added the master Python script used to gather historic data, and the HQL code
that we use to gather article creation statistics.