Optimize new account article creation gathering
We saw an opportunity to improve the performance by switching to batch
processing for identifying article creations and survival. Lookup of
creations is done for a whole batch and is efficient since the revision
ID column is indexed. A temporary table is used to store information about
identified creations and then used to find all deleted articles in a single
query with a join against the logging table.