Rename streamer, update NPP snapshot script
Moved the script that grabs predictions for newly created articles to
a subdirectory where it can live together with other article quality
prediction-related programs. Updated the NPP snapshot script so it
updates the "npp_queue_size" table in our project database on the
enwiki server, where the data can be queries from Quarry.