Update NPP snapshot script for upcoming c1/c3 decomission
The c1/c3 DB servers will be decomissioned, and with them the ability
to create user databases at will. In preparation for that, we moved
the NPP snapshot table back to our database on tools.labsdb and
reverted to the initial code version so data gets written there.

At the same time we also updated the enwiki DB server hostname so
it connects to the new analytics-oriented DB server.