Fix patrol action counts and add summary view.
Because we used a full join to calculate duplicate patrols, we would
miss a lot of patrols. Switching to a left join fixes that problem.
Also added a view that gives the summary statistics per day based
on number of patrol actions, so we can easily find out how many
patrollers were active per day, and the total number of patrol actions.