Add support for manually entered NPP backlog data
This commit documents the changes done to the NPP backlog table
and data gathering in order to support having manually logged
data in the table. We add a new column "npp_method" to the table,
and the column can have one of two values: "automatic" is used for
data points logged by our NPP backlog logging script, while "manual"
is for data points gathered manually. The default value is "automatic".

The backlog logging script is updated to support this change. Lastly,
a SQL file with insert statements to add manually logged data from
https://en.wikipedia.org/wiki/Wikipedia_talk:New_pages_patrol/Reviewers/Backlog_chart_2
This SQL file ignores the first data point as that is unreliable.
The remaining data points go up until 2017-08-29 when our
automatically gathered data points take over.